add a pausemanager field which handles pausing, resuming and managing game-related UI actions such as returning to the main menu or restarting the level.
PauseManager is initialized in the constructor to provide callbacks for pausing the game, resuming the game, exiting to the main menu and restarting the current level.

onPause() and onResume() method are used to handle pausing and resuming.

exitToMainMenu() method is used to handle the transition to the main menu.

restartLevel() method is used to restart the current level.

notifyExitToMainMenu() method  and notifyRestartLevel() method are used to notify LevelListener objects about actions.

Integrated PauseManager to manage pausing via the ESC key.