change background image
decrease target number of kills to win the level to 5.

Keep initializeFriendlyUnits in LevelOne empty to ensure that additional units or custom logic can be added in child classes without altering the core functionality in the parent or GameInitializer.

remove newEnemy.setSpeed() because It does nothing functionally as i set the level one enemy at normal speed